refactor: Update import paths for PaymentStatus type across multiple test and source files
This commit refactors the import statements for the `PaymentStatus` type, changing the path from relative imports to a unified import from "@/types". This change enhances code consistency and maintainability across the codebase, ensuring that all components and tests reference the type from a single source.
